requirements.txt
asarPy==1.0.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
fonttools==4.58.4
fsspec==2024.6.1
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numpy==2.3.0
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
setuptools==70.2.0
six==1.17.0
sympy==1.13.3
torch==2.7.1+cpu
torchvision==0.22.1+cpu
typing_extensions==4.12.2
